build-agent: /chore: move issueClassifier from triggers to core
